pyproject.toml
[build-system]
requires = ["setuptools>=70.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "keypebble"
version = "0.1.2"
description = "Lightweight token issuing service for local, edge, and development environments."
readme = "README.md"
license = "Apache-2.0"
authors = [{ name = "Matt C." }]
requires-python = ">=3.11"
keywords = ["authentication", "jwt", "flask", "security", "edge", "token"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Operating System :: OS Independent"
]
dependencies = [
  "flask>=3.0",
  "pyjwt>=2.8",
  "pyyaml>=6.0",
  "prometheus_client>=0.20"
]

[project.optional-dependencies]
dev = [
  "pytest",
  "requests",
  "build",
  "black",
  "ruff",
  "validate-pyproject",
  "toml-sort",
  "pre-commit"
]

[project.scripts]
keypebble = "keypebble.cli:main"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ["py311"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 88
target-version = "py311"
extend-exclude = ["build", "dist", ".venv"]

[tool.ruff.lint]
# Enable common rule families
select = ["E", "F", "I"]   # Errors, pyflakes, import sorting
ignore = ["E501"]          # Let Black handle line length

[tool.ruff.format]
# Keep compatible with Black
quote-style = "double"
indent-style = "space"
